refactor(detail): 상세페이지 이미지/정보 컴포넌트 분리 및 구조 개선 (#37)
